feat(calendar): extract state to separate primitive
